refactor graph structure, add generic node in api dep graph
