+++ rapx/src/analysis/api_dep/mod.rs
@@ -1,6 +1,4 @@
 mod graph;
-mod lifetime;
-mod ty;
 mod visitor;
 use crate::{rap_info, utils::fs::rap_create_file};
 use graph::ApiDepGraph;
